solve the linux Cmake warning issue

+++ src/tools.cpp
@@ -30,7 +30,7 @@ VectorXd Tools::CalculateRMSE(const vector<VectorXd> &estimations,
     return rmse;
   }
 
-  for (int i=0; i < estimations.size(); ++i)
+  for (size_t i=0; i < estimations.size(); ++i)
   {
     VectorXd error = estimations[i] - ground_truth[i];
     error = error.array()*error.array();
